+++ cpp/subprojects/boosting/include/mlrl/boosting/rule_evaluation/rule_evaluation_non_decomposable_partial_fixed_binned.hpp
@@ -26,9 +26,9 @@ namespace boosting {
 
             const uint32 maxLabels_;
 
-            const float64 l1RegularizationWeight_;
+            const float32 l1RegularizationWeight_;
 
-            const float64 l2RegularizationWeight_;
+            const float32 l2RegularizationWeight_;
 
             const std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr_;
 
@@ -60,8 +60,8 @@ namespace boosting {
              *                                  routines
              */
             NonDecomposableFixedPartialBinnedRuleEvaluationFactory(
-              float32 labelRatio, uint32 minLabels, uint32 maxLabels, float64 l1RegularizationWeight,
-              float64 l2RegularizationWeight, std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr,
+              float32 labelRatio, uint32 minLabels, uint32 maxLabels, float32 l1RegularizationWeight,
+              float32 l2RegularizationWeight, std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr,
               const Blas& blas, const Lapack& lapack);
 
             std::unique_ptr<IRuleEvaluation<DenseNonDecomposableStatisticVector<float64>>> create(
