+++ cpp/subprojects/boosting/src/mlrl/boosting/binning/label_binning_equal_width.cpp
@@ -202,8 +202,8 @@ namespace boosting {
 
     std::unique_ptr<IDecomposableRuleEvaluationFactory>
       EqualWidthLabelBinningConfig::createDecomposableCompleteRuleEvaluationFactory() const {
-        float64 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
-        float64 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
+        float32 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
+        float32 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
         std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr =
           std::make_unique<EqualWidthLabelBinningFactory>(binRatio_, minBins_, maxBins_);
         return std::make_unique<DecomposableCompleteBinnedRuleEvaluationFactory>(
@@ -214,8 +214,8 @@ namespace boosting {
       EqualWidthLabelBinningConfig::createDecomposableFixedPartialRuleEvaluationFactory(float32 outputRatio,
                                                                                         uint32 minOutputs,
                                                                                         uint32 maxOutputs) const {
-        float64 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
-        float64 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
+        float32 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
+        float32 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
         std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr =
           std::make_unique<EqualWidthLabelBinningFactory>(binRatio_, minBins_, maxBins_);
         return std::make_unique<DecomposableFixedPartialBinnedRuleEvaluationFactory>(
@@ -226,8 +226,8 @@ namespace boosting {
     std::unique_ptr<ISparseDecomposableRuleEvaluationFactory>
       EqualWidthLabelBinningConfig::createDecomposableDynamicPartialRuleEvaluationFactory(float32 threshold,
                                                                                           float32 exponent) const {
-        float64 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
-        float64 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
+        float32 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
+        float32 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
         std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr =
           std::make_unique<EqualWidthLabelBinningFactory>(binRatio_, minBins_, maxBins_);
         return std::make_unique<DecomposableDynamicPartialBinnedRuleEvaluationFactory>(
@@ -237,8 +237,8 @@ namespace boosting {
     std::unique_ptr<INonDecomposableRuleEvaluationFactory>
       EqualWidthLabelBinningConfig::createNonDecomposableCompleteRuleEvaluationFactory(const Blas& blas,
                                                                                        const Lapack& lapack) const {
-        float64 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
-        float64 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
+        float32 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
+        float32 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
         std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr =
           std::make_unique<EqualWidthLabelBinningFactory>(binRatio_, minBins_, maxBins_);
         return std::make_unique<NonDecomposableCompleteBinnedRuleEvaluationFactory>(
@@ -248,8 +248,8 @@ namespace boosting {
     std::unique_ptr<INonDecomposableRuleEvaluationFactory>
       EqualWidthLabelBinningConfig::createNonDecomposableFixedPartialRuleEvaluationFactory(
         float32 outputRatio, uint32 minOutputs, uint32 maxOutputs, const Blas& blas, const Lapack& lapack) const {
-        float64 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
-        float64 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
+        float32 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
+        float32 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
         std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr =
           std::make_unique<EqualWidthLabelBinningFactory>(binRatio_, minBins_, maxBins_);
         return std::make_unique<NonDecomposableFixedPartialBinnedRuleEvaluationFactory>(
@@ -260,8 +260,8 @@ namespace boosting {
     std::unique_ptr<INonDecomposableRuleEvaluationFactory>
       EqualWidthLabelBinningConfig::createNonDecomposableDynamicPartialRuleEvaluationFactory(
         float32 threshold, float32 exponent, const Blas& blas, const Lapack& lapack) const {
-        float64 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
-        float64 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
+        float32 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
+        float32 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
         std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr =
           std::make_unique<EqualWidthLabelBinningFactory>(binRatio_, minBins_, maxBins_);
         return std::make_unique<NonDecomposableDynamicPartialBinnedRuleEvaluationFactory>(
